+++ src/app/(app)/dashboard/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function Dashboard() {
+  return <>대시보드</>;
+}
