+++ src/app/(app)/layout.tsx
@@ -1,7 +1,5 @@
-export default function PublicLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
+import AppShell from '@/src/components/layout/AppShell';
+
+export default function AppLayout({ children }: { children: React.ReactNode }) {
+  return <AppShell>{children}</AppShell>;
 }
