--- src/app/(public)/dashboard/[id]/page.tsx
@@ -1,3 +1,0 @@
-export default function Home() {
-  return null;
-}
