+++ src/app/(public)/layout.tsx
@@ -1,5 +1,7 @@
-import AppShell from '@/src/components/layout/AppShell';
-
-export default function AppLayout({ children }: { children: React.ReactNode }) {
-  return <AppShell>{children}</AppShell>;
+export default function PublicLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
 }
